+++ packages/firebase_remote_config/ios/Classes/FirebaseRemoteConfigPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface FirebaseRemoteConfigPlugin : NSObject<FlutterPlugin>
+@end