+++ packages/image_picker/android/src/main/java/io/flutter/plugins/imagepicker/ImagePickerPlugin.java
@@ -5,6 +5,7 @@
 package io.flutter.plugins.imagepicker;
 
 import android.os.Environment;
+import android.support.annotation.VisibleForTesting;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry;
@@ -36,7 +37,8 @@ public class ImagePickerPlugin implements MethodChannel.MethodCallHandler {
     channel.setMethodCallHandler(instance);
   }
 
-  private ImagePickerPlugin(PluginRegistry.Registrar registrar, ImagePickerDelegate delegate) {
+  @VisibleForTesting
+  ImagePickerPlugin(PluginRegistry.Registrar registrar, ImagePickerDelegate delegate) {
     this.registrar = registrar;
     this.delegate = delegate;
   }