+++ packages/firebase_messaging/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.5
+
+* Fixed Dart 2 type error.
+
+## 0.2.4
+
+* Updated Google Play Services dependencies to version 15.0.0.
+
 ## 0.2.3
 
 * Updated package channel name